python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.82.5",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
]